PID registers completely re-done, to have 2 full PIDs on both channels
Finally! it took me one week, but then I discovered why it wasn't working... I was just loading the wrong fpga because the pyrpl code wasn't warning me that the path I gave it wasn't correct... Thank you

I had to remove a lot of features (most importantly, the generic filter) to double the PIDs. Also, the registers are completely different now, we'll use the modified pyrpl to control it from now